requirements.txt
pyTelegramBotAPI
psycopg2-binary
python-dotenv
Flask
gunicorn
sshtunnel
paramiko==2.12.0
cryptography==38.0.4